Copy with cat to a folder, Ls show count of folders, files and size of all files, Find save result of names into a file Mv can move a file to a folder

+++ Shell/Class/Functions/CatFunction.cs
@@ -39,7 +39,19 @@ namespace Shell.Class.Functions
             {
                 if (pathSource != "" && pathDest != "")
                 {
-                    File.Copy(pathSource, pathDest, true);
+                    if (Directory.Exists(pathDest))
+                    {
+                        File.Copy(pathSource, pathDest + "/" + Path.GetFileName(pathSource), true);
+                    }
+                    else if (pathDest.Replace("/", "\\")[pathDest.Length - 1] == '\\')
+                    {
+                        Directory.CreateDirectory(pathDest);
+                        File.Copy(pathSource, pathDest + "/" + Path.GetFileName(pathSource), true);
+                    }
+                    else
+                    {
+                        File.Copy(pathSource, pathDest, true);
+                    }
                 }
                 else if (creatDateV == "" && modifDateV == "")
                 {
